Increase health upon pickup potion

+++ Assets/Scripts/Item/Config/HealthPotionConfig.cs
@@ -5,6 +5,6 @@ public class HealthPotionConfig : PotionConfig
 {
 
     public int Amount;
-    public int Interval;
+    public float Interval;
 
 }
